+++ src/entities/OrderListEntity.ts
@@ -1,0 +1,6 @@
+import { OrderEntity } from "./OrderEntity";
+
+export interface OrderListEntity {
+  orders: OrderEntity[],
+  total: number,
+}