+++ src/navigations/MainNavigator.tsx
@@ -1,11 +1,11 @@
-import React from "react";
-import { createStackNavigator } from "@react-navigation/stack";
-import PurchaseNavigator from "./PurchaseNavigator";
-import HomeNavigator from "./HomeNavigator";
-import { UserDataProvider } from "../provider/UserDataProvider";
+import React from 'react'
+import { createStackNavigator } from '@react-navigation/stack'
+import PurchaseNavigator from './PurchaseNavigator'
+import HomeNavigator from './HomeNavigator'
+import { UserDataProvider } from '../provider/UserDataProvider'
 
 const MainNavigator: React.FC = () => {
-  const Stack = createStackNavigator();
+  const Stack = createStackNavigator()
   return (
     <UserDataProvider>
       <Stack.Navigator
@@ -13,13 +13,14 @@ const MainNavigator: React.FC = () => {
         screenOptions={{
           headerShown: false,
           headerBackTitleVisible: false,
+          gestureEnabled: false,
         }}
       >
         <Stack.Screen name="Home" component={HomeNavigator} />
         <Stack.Screen name="Purchase" component={PurchaseNavigator} />
       </Stack.Navigator>
     </UserDataProvider>
-  );
-};
+  )
+}
 
-export default MainNavigator;
+export default MainNavigator